feature: separate GqbIntermediary from BlasterDataGenerator
